requirements.txt
fastapi
uvicorn[standard]
python-multipart
pymupdf
google-generativeai
python-dotenv
sqlalchemy
pydantic[email]
psycopg2-binary
alembic
bcrypt
python-jose[cryptography]
passlib[bcrypt]
